+++ app/Http/Controllers/Auth/LoginController.php
@@ -36,20 +36,6 @@ class LoginController extends Controller
     use AuthenticatesUsers;
 
     /**
-     * Maximum number of attempts to allow.
-     *
-     * @var int
-     */
-    protected $maxAttempts = 5;
-
-    /**
-     * Number of minutes to throttle for.
-     *
-     * @var int
-     */
-    protected $decayMinutes = 1;
-
-    /**
      * Where to redirect users after login.
      *
      * @var string
@@ -64,6 +50,8 @@ class LoginController extends Controller
     public function __construct()
     {
         $this->middleware('guest')->except('logout');
+        $this->maxAttempts = config('auth.login.max_attempts', 5);
+        $this->decayMinutes = config('auth.login.decay_minutes', 1);
     }
 
     /**