+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/) and this 
 ## Unreleased
 ### Added
 - Refactor code to add more tests to improve code coverage.
+- Add two values to `config/auth.php` to configure Login throttling: 'auth.login.max_attempts' and 'auth.login.decay_minutes'.
 
 ## 0.6.0
 > **Note**: This application is now using [Laravel 6](https://laravel.com/docs).